+++ lib/src/flutter/models/cloud_storage.dart
@@ -16,7 +16,7 @@ class CloudStorage {
   Future<void> setItem(String key, String value,
           [void Function(String? error, [bool? isStored])? callback]) =>
       telegram_js.CloudStorage.setItem(
-          key, value, callback != null ? telegram_js.JsDynamicCallback(callback) : null);
+          key, value, callback != null ? JsDynamicCallback(callback) : null);
 
   /// A method that receives a value from the cloud storage using the specified key.
   /// [key] should contain 1-128 characters, only A-Z, a-z, 0-9, _ and - are allowed
@@ -24,7 +24,7 @@ class CloudStorage {
   ///  will contain the error. In case of [success], the first argument will be null and
   ///  the value will be passed as the second argument.
   Future<void> getItem(String key, void Function(String? error, [String? result]) callback) =>
-      telegram_js.CloudStorage.getItem(key, telegram_js.JsDynamicCallback(callback));
+      telegram_js.CloudStorage.getItem(key, JsDynamicCallback(callback));
 
   /// A method that receives values from the cloud storage using the specified keys.
   /// [keys] should contain 1-128 characters, only A-Z, a-z, 0-9, _ and - are allowed.
@@ -33,7 +33,7 @@ class CloudStorage {
   /// and the values will be passed as the second argument.
   Future<void> getItems(
       List<String> keys, void Function(String? error, [List<String>? values]) callback) {
-    return telegram_js.CloudStorage.getItems(keys, telegram_js.JsDynamicCallback(callback));
+    return telegram_js.CloudStorage.getItems(keys, JsDynamicCallback(callback));
   }
 
   /// A method that removes a value from the cloud storage using the specified key.
@@ -44,7 +44,7 @@ class CloudStorage {
   /// whether the value was removed.
   Future<void> removeItem(String key, [void Function(String? error, [bool? removed])? callback]) =>
       telegram_js.CloudStorage.removeItem(
-          key, callback != null ? telegram_js.JsDynamicCallback(callback) : null);
+          key, callback != null ? JsDynamicCallback(callback) : null);
 
   /// A method that removes values from the cloud storage using the specified keys.
   /// [keys] should contain 1-128 characters, only A-Z, a-z, 0-9, _ and - are allowed.
@@ -55,13 +55,12 @@ class CloudStorage {
   Future<void> removeItems(List<String> keys,
           [void Function(String? error, [bool? removed])? callback]) =>
       telegram_js.CloudStorage.removeItems(
-          keys, callback != null ? telegram_js.JsDynamicCallback(callback) : null);
+          keys, callback != null ? JsDynamicCallback(callback) : null);
 
   /// A method that receives the list of all keys stored in the cloud storage.
   /// In case of an [error], the callback function will be called and the first argument will contain
   /// the error. In case of [success], the first argument will be null and the list of keys will be
   /// passed as the second argument.
   Future<void> getKeys(void Function(String? error, [List<String>? keys]) callback) =>
-      telegram_js.CloudStorage.getKeys(
-          telegram_js.JsDynamicCallback(telegram_js.JsDynamicCallback(callback)));
+      telegram_js.CloudStorage.getKeys(JsDynamicCallback(callback));
 }