+++ lib/src/js/telegram/src/models/biometric/biometric_manager.dart
@@ -9,8 +9,8 @@ class BiometricManager {
   external bool get isAccessGranted;
   external bool get isBiometricTokenSaved;
   external String get deviceId;
-  external Future<void> init([callback]);
 
+  external Future<void> init([callback]);
   external Future<void> requestAccess(BiometricRequestAccessParams params, [callback]);
   external Future<void> authenticate(BiometricAuthenticateParams params, [callback]);
   external Future<void> updateBiometricToken(token, [callback]);