+++ lib/src/js/telegram/telegram_js_models.dart
@@ -5,6 +5,8 @@ library telegram_js_internal;
 
 import 'package:js/js.dart';
 
+import '../telegram_js.dart';
+
 part 'src/events/write_access_requested_event_result.dart';
 part 'src/models/back_button.dart';
 part 'src/models/biometric/biometric_authenticate_params.dart';