+++ lib/src/telegram_app.dart
@@ -74,6 +74,9 @@ abstract class TelegramWebApp {
   /// An object for controlling cloud storage.
   CloudStorage get cloudStorage;
 
+  /// An object for controlling biometrics on the device.
+  BiometricManager get biometricManager;
+
   /// Returns true if the user's app supports a
   /// version of the Bot API that is equal to or
   /// higher than the version passed as the parameter.